requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
statsmodels
pmdarima
prophet
tensorflow-cpu
keras
jupyter